fixed file exists bug

+++ get_small_dim_images.py
@@ -7,14 +7,14 @@ from tqdm import tqdm #progress bar
 import sys
 
 
-def _traverse_folders(root_folder,output_folder):
+def _traverse_folders(root_folder,output_folder,wanted_files):
 	'''
 	function to traverse a folder structure and run a function in each folder
 	'''
 	for folder in tqdm(os.listdir(root_folder)):
 		print("Processing: {}".format(folder))
 		if folder.startswith("MS-DAR-"):
-			_copy_jpg_on_dim(folder,output_folder)
+			_copy_jpg_on_filename(folder,output_folder,wanted_files)
 
 def _traverse_folders_write(root_folder):
 	'''
@@ -45,8 +45,7 @@ def _copy_jpg_on_dim(root_folder,output_folder):
 
 def _copy_jpg_on_filename(root_folder,output_folder,wanted_files_txt):
 	'''
-	function to 1) get average dimensions for all images in a folder
-				2) copy all images less than average to a target folder
+	function to 1) copy all images from a textfile of desired images
 	'''
 	wanted_files = [line.rstrip('\n') for line in open(wanted_files_txt)]
 	if root_folder.startswith("MS-DAR-"):
@@ -55,11 +54,12 @@ def _copy_jpg_on_filename(root_folder,output_folder,wanted_files_txt):
 		prev_num = 0
 		for file in tqdm(os.listdir(root_folder)):
 			if file in wanted_files:
-				copyfile(os.path.join(root_folder,file),os.path.join(output_folder,file))
+				if not os.path.exists(os.path.join(output_folder,file)):
+					copyfile(os.path.join(root_folder,file),os.path.join(output_folder,file))
 
-def _traverse_folders_spark(root_folder,output_folder,wanted_files="all_cut_jpg_filenames.txt"):
+def _traverse_folders_spark(root_folder,output_folder,wanted_files):
 	'''
-	function to traverse a folder structure and run a function in each folder
+	spark function to traverse a folder structure and run a function in each folder
 	'''
 	rootFolderRDD = sc.parallelize(os.listdir(root_folder))
 	#placeholderRDD = rootFolderRDD.map(lambda x: _copy_jpg_on_dim(x,output_folder)).collect()
@@ -69,4 +69,4 @@ if __name__ == "__main__":
 	from pyspark import SparkContext
 	sc = SparkContext()
 	print("Executing as main program")
-	_traverse_folders_spark(sys.argv[1],sys.argv[2],sys.arg[3])
+	_traverse_folders_spark(sys.argv[1],sys.argv[2],sys.argv[3])
